Updated CurrencyInput.
