+++ packages/components/src/CurrencyInput/types.ts
@@ -1,7 +1,6 @@
 export interface CurrencyInputProps {
 	amount: string | number;
-	id: string;
-	onChange?: (result?: { amount: string | number; id: string }) => void;
+	input: (formattedAmount: string) => React.ReactNode;
 	placeholder?: string;
 	tag?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'p';
 	tooltip?: string;
